// src/types/cart.d.ts
export interface ExchangeRate {
    table: string
    rates: Rates
    lastupdate: string
}

export interface Rates {
    AED: number
    AFN: number
    ALL: number
    AMD: number
    ANG: number
    AOA: number
    ARS: number
    AUD: number
    AWG: number
    AZN: number
    BAM: number
    BBD: number
    BDT: number
    BGN: number
    BHD: number
    BIF: number
    BMD: number
    BND: number
    BOB: number
    BRL: number
    BSD: number
    BTC: number
    BTN: number
    BTS: number
    BWP: number
    BYN: number
    BZD: number
    CAD: number
    CDF: number
    CHF: number
    CLF: number
    CLP: number
    CNH: number
    CNY: number
    COP: number
    CRC: number
    CUC: number
    CUP: number
    CVE: number
    CZK: number
    DASH: number
    DJF: number
    DKK: number
    DOGE: number
    DOP: number
    DZD: number
    EGP: number
    ERN: number
    ETB: number
    ETH: number
    EUR: number
    FJD: number
    FKP: number
    GBP: number
    GEL: number
    GGP: number
    GHS: number
    GIP: number
    GMD: number
    GNF: number
    GTQ: number
    GYD: number
    HKD: number
    HNL: number
    HRK: number
    HTG: number
    HUF: number
    IDR: number
    ILS: number
    IMP: number
    INR: number
    IQD: number
    IRR: number
    ISK: number
    JEP: number
    JMD: number
    JOD: number
    JPY: number
    KES: number
    KGS: number
    KHR: number
    KMF: number
    KPW: number
    KRW: number
    KWD: number
    KYD: number
    KZT: number
    LAK: number
    LBP: number
    LD: number
    LKR: number
    LRD: number
    LSL: number
    LTC: number
    LYD: number
    MAD: number
    MDL: number
    MGA: number
    MKD: number
    MMK: number
    MNT: number
    MOP: number
    MRO: number
    MRU: number
    MUR: number
    MVR: number
    MWK: number
    MXN: number
    MYR: number
    MZN: number
    NAD: number
    NGN: number
    NIO: number
    NOK: number
    NPR: number
    NXT: number
    NZD: number
    OMR: number
    PAB: number
    PEN: number
    PGK: number
    PHP: number
    PKR: number
    PLN: number
    PYG: number
    QAR: number
    RON: number
    RSD: number
    RUB: number
    RWF: number
    SAR: number
    SBD: number
    SCR: number
    SDG: number
    SEK: number
    SGD: number
    SHP: number
    SLL: number
    SOS: number
    SRD: number
    SSP: number
    STD: number
    STN: number
    STR: number
    SVC: number
    SYP: number
    SZL: number
    THB: number
    TJS: number
    TMT: number
    TND: number
    TOP: number
    TRY: number
    TTD: number
    TWD: number
    TZS: number
    UAH: number
    UGX: number
    USD: number
    UYU: number
    UZS: number
    VEF_BLKMKT: number
    VEF_DICOM: number
    VEF_DIPRO: number
    VES: number
    VND: number
    VUV: number
    WST: number
    XAF: number
    XAG: number
    XAU: number
    XCD: number
    XDR: number
    XMR: number
    XOF: number
    XPD: number
    XPF: number
    XPT: number
    XRP: number
    YER: number
    ZAR: number
    ZMW: number
    ZWL: number
    NMC: number
    PPC: number
    NVC: number
    XPM: number
    EAC: number
    VTC: number
    EMC: number
    FCT: number,
    [x: string]: any
}

export interface CartItems {
    _id: string
    userId: string
    items: Item[]
    totalAmount: number
    __v: number
}

export interface CartItem {
    productId: string
    itemId?: string
    productName: string
    productType: string
    description?: string
    productImages: string[]
    productQuantity: number
    pricePerUnit: number
    currency: string
    storeId?: string
    storeName: string
    size: string
    color: string
    purchaseQuantity: number
    purchaseAmount: number
}

export interface Product {
    productId: string
    productName: string
    productType: string
    productImages: any[]
    currency: string
    storeName: string
    size: any[]
    color: any[]
}